Remove some more code that isn't necessary
